interaction readability, expandability and refactor
