+++ Cargo.toml
@@ -21,3 +21,4 @@ regex = "1.10.2"
 octocrab = "0.19.0"
 reqwest = "0.11.22"
 hex = "0.4.3"
+to-arraystring = "0.1.3"
